--story=1000045 --user=郑富城 C++全面优化 - BasePixelCharacter 基类 https://www.tapd.cn/42563989/s/1000099 # BaseWeapon

+++ Pixel2DKit/Public/Basic/PXGameState.h
@@ -16,7 +16,7 @@ struct FLevelWeatherRate: public FTableRowBase
 	TMap<FName, int> WeatherRate;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Weather")
-	TMap<float, float> TimeOfDayRate;
+	TMap<float, int> TimeOfDayRate;
 };
 
 USTRUCT(BlueprintType)
